Revert "[refactor] Server Status API から ACL クラスを分離, 単体テストの作成 (#53)"
This reverts commit 8103de3f89839913fa7a1e090647a0006ccc89ca.
